refactor(temporal): create specific testUnitStruct for test group mpe
